:recycle: Adicionado teste para busca por identificador e atualização por referencia
