+++ cap2/ecommerce/src/test/java/com/lucasengcomp/ecommerce/iniciandocomjpa/EntityManagerTest.java
@@ -1,0 +1,37 @@
+package com.lucasengcomp.ecommerce.iniciandocomjpa;
+
+import org.junit.After;
+import org.junit.AfterClass;
+import org.junit.Before;
+import org.junit.BeforeClass;
+
+import javax.persistence.EntityManager;
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.Persistence;
+
+public class EntityManagerTest {
+
+    protected static EntityManagerFactory entityManagerFactory;
+    protected EntityManager entityManager;
+
+    @BeforeClass
+    public static void setUpBeforeClass() {
+        entityManagerFactory = Persistence
+                .createEntityManagerFactory("Ecommerce-PU");
+    }
+
+    @AfterClass
+    public static void tearDownAfterClass() {
+        entityManagerFactory.close();
+    }
+
+    @Before
+    public void setUp() {
+        entityManager = entityManagerFactory.createEntityManager();
+    }
+
+    @After
+    public void tearDown() {
+        entityManager.close();
+    }
+}
